fix the wrong write_to_swap

+++ src/vm/swap.h
@@ -11,7 +11,7 @@ void read_from_block(void* frame, int index);
 size_t write_to_swap(void *kpage);
 size_t read_from_swap(void *kpage, size_t start);
 struct bitmap *swap_bmap;
-int swap_index_global;
+size_t swap_index_global;
 static struct block *swap_block;
 struct lock swap_lock;
